added the command line option `--print-gaproot`
Note that the list of command line options in the GAP Reference Manual
(Section "Command Line Options") does not contain several of the
options that have been added recently.
More precisely, this list does not mention options with "long form"
that start with `--`, and does not list any of these options.
Should we extend the list in the manual,
or should we perhaps better refer to the list printed by the `-h`
(or `--help`) option?

resolves #4448