ENHANCE: Default representation for perfect groups
The default choice of representation used by the perfect groups library is
as finitely presented group. (This was decided by V. Felsch, when he
implemented this library in the 1990s.)

However, short of constructing or storing, this representation is almost
universally inferior to a permutation representation. This commit thus
switches this default.