Implement IsMinimalNonmonomial for nonsolvable grp
The previous implementation followed van der Waall in only handling
solvable groups. Indeed in the paper,
[MR507832](https://mathscinet.ams.org/mathscinet-getitem?mr=507832)
van der Waall says the nonsolvable groups can easily be disposed
of using Thompson's N-groups papers, however he does not state the
result.

Luckily most of the minimal simple groups are also minimal
non-monomial. The exception is just PSL(3,3) which contains a subgroup
isomorphic to the non-monomial group SL(2,3). The groups of the form
PSL(2,q) are handled by Dickson's classification of their subgroups,
all of which are metabelian or non-solvable. Since the non-solvable
subgroups are eliminated by being minimal simple, we have only
monomial subgroups. The Suzuki groups require a short argument,
given in the source. Suzuki 1960 lists the maximal subgroups, and all
but the Sylow 2-normalizer are metabelian. The Sylow 2-normalizer's
characters are described in sufficient detail in the paper to ensure
they are monomial.

I included a commented out "IsMNMNaive" that checks all the
conditions. I ran this for 10 or 20 minutes on the small, perfect,
and primitive group libraries.