kernel: better keep track of stdout/errout state
The GAP kernel keeps track of the current output device for printing via
a stack of output "files", represented by structs of type
`TypOutputFile`. For each open file, GAP keeps track of its state (such
as the current cursor position, in such a struct.

However, the standard output is somewhat unique here in that it may be
opened multiple times (both as "stdout" and also as "errout"). But so
far GAP did never synchronize the state between the potentially multiple
structs tracking the state of stdout/errout. This leads to weird
printing effects, such as GAP printing line continuations in unexpected
places (too later or too early). This was exacerbated by the fact that
printing to errout in `lib/error.g` happens via calls to
`PrintTo(ERROR_OUTPUT, ...)` which actually opens and then immediately
again closes errout; so tracking any information on cursor position or
hints for word wrapping would not be carried forward correctly.

Moreover, flushing the output buffer via `\c` incorrectly reset the
presumed cursor position to 0; this was because the same variable `pos`
was used to track the cursor position, and the current offset into the
line buffer.

This patch fixes all that, by doing two things: First, whenever we open
or close a `TypOutputFile` referencing stdout or errout, we synchronize
it with the previous `TypOutputFile` referencing stdout or errout, if
any. Secondly, we add a new field to `TypOutputFile` which keeps track
of the cursor position separately from the line buffer state.

This revealed a bug where parts of the GAP banner suddenly were wrapped
unexpectedly. This in turn is caused by us printing some unicode
characters in the banner (to draw a nice ASCII art box around the word
GAP); but since these unicode characters each take up more than one
byte, this threw of our cursor tracking code. The reason this worked
before and broke now is that the code printing the box actually abused
the fact that flushing the line buffer by printing `\c` also reset the
tracked cursor position. But we just fixed this... So as a new
workaround, we temporarily disable output formatting for stdout.