Replace GAPDoc's PrintFormattedString by PrintWithoutFormatting
`PrintWithoutFormatting` is a new helper function which calls
`(Set)PrintFormattingStatus("*current*")` suitably to temporarily
modify the formatting status for the currently active output stream.

This requires adjusting one test for `Info`, which printed a notice
about `InfoGlobal` which was in turn caused by a call to `IsBoundGlobal`
removed by this patch.