use the new definition of `IsMatrixObj` in methods
In each method installation that takes an argument in `IsMatrixOrMatrixObj`
such that the function exits with `TryNextMethod()` if this argument is in
`IsPlistRep`,
restrict the requirement to `IsMatrixObj` and omit the check for `IsPlistRep`.

Also removed some ancient (from GAP 3 times) comment lines,
and added a remark to the documentation that `IsRowListMatrix` need not
imply `IsList`.