Add `InstallEarlyMethod`
This installs a special "early" function method for a given operation.
An early method is special in that it bypasses method dispatch, and is
always the first method to be called when invoking the operation. This
can be used to avoid method selection overhead for certain special
cases.

Unlike with regular methods, no checks are performed on the arguments.
Hence any operation can have at most one such early method for each
arity (i.e., one early method taking 1 argument, one early method taking
2 arguments, etc.).

If an early method detects that it is not applicable, it can resume
regular method dispatch by invoking `TryNextMethod`.

For example, we install early methods for `First` that deal with
internal lists, for which computing their types (and hence method
selection) can be very expensive.