explicitly document that `IsRowListMatrix` objects are "dense"
The documentation of the relevant operations for objects in `IsRowListMatrix`
(list access, list assignment, `IsBound[]`, ...) implies that such objects are
"dense" (although they need not be lists),
and this follows also from the fact that such an object `M` is in `IsMatrixObj`
and hence admits entry access `M[i,j]`,
for `i`and `j` up to `NumberRows( M )` and `NumberColumns( M )`, respectively.
However, it is better to say this explicitly (see also #4533).