Revise our pull request template
The old one was far too long, meaning that most people would skip
over most of it. So let's focus on the important bits, which are
a useful description and release notes.

The stuff at the end for reviewers was not used by us in practice;
instead, there should be a separate document somewhere with guidelines
for reviewers.