First version of a new HelpLinks to JSON script
This is based on, and replaces, the `etc/LinksOfAllHelpSections.g`
script of the GapWWW repository.

Instead of this script being run over there, on the newly released
version of GAP, I think it instead makes sense to create this file
while making all of the other archives, and including it in the
GitHub release.