PERFECT: Reduced the permutation degree for a few groups (#4573)
The library stores groups through a presentation with words for the subgroup
stabilizer(s). In a few cases the groups created had an unreasonably large
permutation degree. This patch reduces the degrees by changing the
stabilizers. The ptresentations (and thus the groups) are not affected,
reducing the potential for accidental errors.

None of this affects the groups in the ``classical'' perfect groups library