kernel: get rid of CATCH_ERROR macro
It turns out all cases where it was used are now redundant: they were
all about calling `IntrAbortCoding`. But since all the interpreter and
coder state is on the stack now, and also `IntrEnd` invokes
`IntrAbortCoding` for us if there was an error, we can just get rid of
all these cases.