Fix ViewString and DisplayString for ranges
... to match ViewObj and Display.

As a result, Display, PrintObj, DisplayString, PrintString and String
now all behave the same for ranges, while ViewObj and ViewString differ.
This is probably quite surprising to people, but changing it one way or
the other will affect packages, so for now let's play it safe and only
address the "obvious" bugs.