FIX: Generic stabilizer if order known
If the group order is known, the generic orbit stabilizer might stop
enumerating the orbit if one can deduce that it could not grow anymore for
divisibility reasons.

In this case the stabilizer in rare cases might still be too small and need
to be corrected. This was done not fully.

Thsi fixes #4659