Enable tab completion for non-internal records
If one implements a non-internal record type, then with this patch tab
completion for instances of this type becomes possible.

Note that this means we can't call `NamesOfComponents` to get the record
names, in case the record type is implemented as a component object:
`NamesOfComponents` will then return the internal names of that
component object, not the names of the record it implements.

For the same reason, we must split the code, which so far uniformly used
`r!.(name)` to access components of record and component objects.

There is other code which arguably should work with non-internal record
objects but currently calls `NamesOfComponents`; I've adjusted a few
more of these to call `RecNames` instead, but this is likely not
comprehensive.

Note that I deliberately decided against changing the behaviour of
`NamesOfComponents` as it is explicitly documented, has been so for some
time, and a bunch of packages use it.

To test this, enter this GAP code:

    fam := NewFamily("MockFamily");;
    cat := NewCategory("IsMockRecord", IsRecord);;
    type := NewType(fam, cat and IsPositionalObjectRep);;
    InstallMethod(RecNames, [cat], r -> ["dummy"] );
    mockRec := Objectify(type,["myName"]);;

Then type `mockRec.` and press TAB. It should complete to `mockRec.dummy`.