buildsys: restrict vpath to src directory
The `vpath` directives are what makes out-of-tree builds possible. With
them, if a source file is not present in `${builddir}`, then GNU make
automatically looks for it in `${srcdir}`.

For files in `src`, this is generally what we want. But for generated
files in `build`, we do *not* want that. As a result, it could happen
that instead of `${builddir}/build/gap_version.c` we would access
`${srcdir}/build/gap_version.c`, and so on.

This patch fixes this by restricting the `vpath` directives to files
inside `src` subdirectory.