FIX: Use `ImageElm` instead of `Image` in `ListWreathProductElement`.
If the top group is not the full symmetric group,
trying to construct the `Image` of an element might result in an error,
since the element might not be in the `Source`.
Using `ImageElm` this might now only cause a `fail`.
This fix is important if we want to use `ListWreathProductElement`
for a fast `\in` check in wreath products.