kernel: improve position tracking after ReadEvalCommand
When evaluating inputs, ReadEvalCommand and the GAP scanner greedily read full
lines of input, even if only a part of the line ends up being evaluated; e.g.
because there are multiple commands on a line.

Unfortunately, this rendered the (undocumented) command `READ_COMMAND_REAL`
far less useful: when invoked on a generic input stream, it evaluates exactly
one command; but afterwards the end of that command can not be inferred from
the position of the input stream, as it always reads full lines.

This patch changes things so that when closing an input *stream*, backed by a
GAP stream object, we try adjust its seek position to where the actual command
ended.

This means that now one can repeatedly call `READ_COMMAND_REAL` on a string
stream to iteratively evaluated the statements in the underlying string.
Combined with `CALL_WITH_STREAM` this mostly allows to reimplement
`READ_ALL_COMMANDS` in GAP code, but with far more flexibility. This is
potentially useful for e.g. the Jupyter interface.

(The one thing that is not yet available is detection of dual semicolons,
which `READ_ALL_COMMANDS` supports but `READ_COMMAND_REAL` does not. But this
could be tacked on to `READ_ALL_COMMANDS`, or we could implement a new
variant of this command providing that information.)