kernel: turn ExecStatus into a proper enum
... and don't use it as a bitmask. It was a bitmask before as a (minor)
performance optimization. Alas, in most call sites, this did not matter: the
read-eval-loops in gap.c and streams.c are not affected by a few extra
instructions, especially if one is careful to first check the common case
(STATUS_END) before the exceptional cases (STATUS_QUIT, STATUS_QQUIT,
STATUS_ERROR, STATUS_EOF)

The only potentially performance critical check was in `EXEC_STAT_IN_LOOP`,
which was already switched away from using the bitmasks in a previous commit.
But even there, really only the case of a `break` or `return` statement
inside of a loop is affected, which once again is an "exceptional" case.