gac: remove -r and last traces of -fFEATURE
The -r option was the only reason for referencing abs_top_srcdir, and nothing
seems to use it. I can't think of any reason for using it either, other than
perhaps convenience when invoking this manually (instead of via a build script).

The -f options like -ffast-int-arith were removed in 2017 without complaints
since then, so we can safely remove the last vestiges of them now.