changed a generic `OneSameMutability` method
Up to now, the lowest rank method for `OneSameMutability` called
`OneMutable` and then `MakeImmutable` for an immutable argument.
This approach runs into a "no method found" error if the argument
is a domain.

Note that the analogous situation for `ZeroImmutable` was different:
Here the lowest rank method calls `ZeroAttr` for an immutable
argument.