workaround for a `One` method
GAP Reference Manual defines `One` only for an argument in `IsMultiplicativeElementWithOne` or `IsDomain`.
A method in `lib/arith.gi` tries to extend this definition to collections of `IsMultiplicativeElementWithOne` objects that aren`t domains, by delegating to a representative of the collection.

This causes a logical problem if the collection is itself in `IsMultiplicativeElementWithOne`, because of a different meaning.
Such a situation occurs in the case of a group character, for which `One` shall return not `1` but the trivial character.

It would be good if we could eventually get rid of this method.
For the moment, the method gets changed in such a way that it delegates
to the next applicable method if the argument is in `IsMultiplicativeElementWithOne`.