kernel: get rid of `name` field in `struct TypInputFile`
This results in a super tiny amount of extra work for reading empty files /
files with only comments but no code. On the up side it opens the way for
cleanly supporting paths longer than 256 chars. We still have limits in other
places, but I think they are mostly 1204 or 4096 chars.