ENHANCE: re-use cached pcgs
It can happen (e.g. when storing level indices) that routines use pcgs that
are not the HomePcgs identically, but the same pces as induced by the
HomePcgs.  The code for checking whether an induced pcgs is already cached
thus tests for equality, not identity with the HomePcgs. The cost of the
extra test is neglegible over computing an induced pcgs.