fixed a bug in character table `Display`
If one wants to show the 2nd `indicator` in `Display`
and if some indicator is stored as unknown
(which happens for a few 2-modular library tables)
then up to now the column alignment was not correct.

Taking that this effect had not been observed up to now,
I think that the fix need not be mentioned in the release notes.