Remove a test for data in the small groups library
This test started to fail with the release of SmallGrp 1.5, as that corrects
the number of groups of order 1024 which was previously misrecorded.

I could have adjusted the test here, but instead I removed it, as SmallGrp
has its own test suite which already contains similar tests, so there really
is no point in repeating these checks here.