buildsys: change JULIA_CPPFLAGS to JULIA_CFLAGS
... and accordingly only add them to our CFLAGS, not CPPFLAGS. These flags
always were advertised as CFLAGS, and this change avoids incompatibilities of
these flags with C++ compilers that reared up again with Julia 1.9.

Also remove some code removing apostrophes from JULIA_*FLAGS for use in shell
scripts; this is not needed anymore as we don't put them into sysinfo.gap
anymore. We need to keep this transformation for JULIA_LIBS, though, as it is
used as in `-Wl,-rpath,'/SOME/PATH'` and this then confuses libtool.