Avoid using InstallValue on non-plain objects
InstallValue and friends are the only reason we need the somewhat
dangerous kernel function `CLONE_OBJ`. It has known issues when used
on e.g. families (which we already try to reject) and types.

If we can eliminate all uses of it on component and positional objects,
then we could potentially replace it by a simpler implementation (which
would reject attempts to use it on such non-plain objects), which would
also help avoid pitfalls in HPC-GAP.