Revise SmallGeneratingSet
A recent change made SmallGeneratingSet give up very quickly in many
cases. This broke the test suites of a bunch of packages, and in
general provided output which users may regard as suboptimal, e.g.
including trivial permutations and duplicate generators.

This patch tries to go for a compromise: we still abort earlier than
we used to, but not *quite* as early, and we try to at least remove
"obvious" redundancies.