Prepare etc/Makefile.gappkg for future gac changes
In the future I hope to replace all use of GNU libtool in gac,
to make life easier for downstream packagers (e.g. Debian, Fedora, ...)
One of the changes will be that object files created by gac will go
back to use .o as extension instead of .lo. This patch prepares
`Makefile.gappkg` for this.

The idea here is that by including this update now, we can start
changing packages to use it *now*, rather than when the other
gac changes are ready.

(BTW, this is *not* strictly necessary, as I am trying hard in
the changed gac to still support the old `Makefile.gappkg`. That
said, any change we can merge now makes the gac PR smaller and
easier to review)