Disabled some = and in-tests
According to the generic methods for equality \= for IsMatrixObj
and IsVectorObj in lib/matobj.gi, equality needs the same
representation filter, base domain, size and entries.

In lib/matobjnz.gi there are methods for comparison with
lists and lists of lists which yield true when just the entries
coincide. This contradicts the generic definition.
The test examples which are commented out in this commit
rely on these questionable methods.

[I'm not completely sure what the definition of \= should be,
but at least the implementation should be consistent between the various
representations of vectors and matrices.]