kernel: change how our GCs represent master pointers
So far, our master pointers are pointers to a pointer to the
data portion of a bag. This patch changes them to instead
point at the start of the bag.

Pointing inside the bag, as was done before, was am
optimization for the common case where one wants to access
the content of a bag. However, at least on x86 and ARM
architectures, there is no difference in practice, because
adding an offset of 16 (GASMAN) resp. 8 (Boehm, Julia) is
essentially free for most machine code opcodes used to
access the data.