kernel: avoid compiler warnings in some kernel extensions
Kernel extensions which don't use compiler options to turn off
certain warnings may see a bunch of them caused by conversion
between different kinds of function pointers, which in general
can indeed indicate a bug, but here is *really* correct. We
silence these for good by first casting to `void *` which is
always legal.

Also, when compiling with C++, disable a `#pragma` we use to
temporarily silence the warning issued by `-Wstrict-prototypes`,
as this warning is not supported in C++ and so trying to disable
it can cause a warning by itself.