SyCountProcessors function for Linux, Windows and macOS
SyCountProcessors() retrieves the number of active logical processors. In case of Linux, calls get_nprocs(). In case of Windows 7 and higher, calls GetActiveProcessorCount(ALL_PROCESSOR_GROUPS). In case of older versions of Windows calls GetSystemInfo with proper arguments. In case of macOS sysctl is called with HW_AVAILCPU flag. In case of other operating systems the value of 4 is returned.