Fix unexpected error in `ConjugacyClassesSubgroups`
This disables an attempt to use cyclic extension instead of lifting for
lattices of smallish groups. The issue is that CE returns data in a
slightly different format than required. This resulted in some commands,
e.g. `ConjugacyClassesSubgroups(SmallGroup(1344,11293))`, to enter a
break loop.

If tomlib is loded it's a mote issue, and thus does not merit further
investigation.

This resolves #4854