Initializing `SyNumProcessors` in runtime with the number of available CPUs for Linux, Windows and macOS (#4962)
* SyCountProcessors function for Linux, Windows and macOS

SyCountProcessors() retrieves the number of active logical processors. In case of Linux, calls get_nprocs(). In case of Windows 7 and higher, calls GetActiveProcessorCount(ALL_PROCESSOR_GROUPS). In case of older versions of Windows calls GetSystemInfo with proper arguments. In case of macOS sysctl is called with HW_AVAILCPU flag. In case of other operating systems the value of 4 is returned.

* removed a plain comment on SyCountProcessors

Co-authored-by: Max Horn <max@quendi.de>

* cpu.c - fixed some dots trouble inside the comment

Co-authored-by: Max Horn <max@quendi.de>

* reformatted code with clang-format

* moved system specific includes from cpu.h to cpu.c

* fixed SyCountProcessors' comment in cpu.h

Got rid of plain rewrite of what the function does, plus removed blank line between the comment and the function.

* removed blank line SyCountProcessors' after the comment in cpu.c

* SyCountProcessors implementation using sysconf()

As sysconf is available on POSIX-compliant systems, there is no need to write code for Linux and macOS separately.

* got rid of one level of #if in cpu.c

* SyCountProcessors: dropped Windows less than 7

* removed NUM_CPUS macro

* no repeating (magic) number 4 in cpu.c

Introduced a static const UInt FALLBACK_CPUS_NUMBER in cpu.c to use it instead of just 4

* seems like FALLBACK_CPUS_NUMBER should be a macro

Co-authored-by: Max Horn <max@quendi.de>