Speed up DoOrbitStabilizerAlgorithmStabsize for matrix groups
Ensure there is a nice monomorphism on the group in which we compute the
stabilizers. This speeds up the following example command from the
Sophus package:

    AreIsomorphicNilpotentLieAlgebras( L7[100], L7[101] );

Before this patch this takes 13 to 18 seconds on my laptop (timings vary
substantially, presumably due to randomization). After this patch,
this is down to 0.8 to 0.9 seconds.