ImmutableMatrix should not convert
between lists of lists vectors/matrices and vector/matrix objects.

The mentioned representations of vectors and matrices are not
interoperable. The current use of ImmutableMatrix would cause mixed
representations and so various bugs in vector-matrix arithmetic.

The constructor functions for classical groups should explicitly call
'Matrix' with the appropriate rings if generators (and matrices of
invariant forms) should be matrix objects.