GAPInfo: add NonBareDependencies item.
If GAP is loaded with --bare, GAPInfo.Dependencies is overwritten with an empty
list.  One obvious use-case of --bare is loading GAP without packages for just
long enough to use PackageManager to install required packages.  PackageManager
has the InstallRequiredPackages function for this purpose, but since it requires
GAPInfo.Dependencies, it ironically becomes useless in the one circumstance
where it's really needed!

This commit would allow the normal required packages to be found in
GAPInfo.NonBareDependencies even if --bare is used, which will allow
PackageManager to install required packages without the need to hard-code
anything.