Revert recent changes to FieldOfMatrixList
The (unofficial, not included in the reference manual) documentation for
`FieldOfMatrixList` states that it returns the "smallest field containing all
the entries of all matrices in [the list given to it]".

The code removed by this patch violated this promise, by using the `BaseDomain`
of the given matrices, which may *not* be the smallest field; this in turn broke
some code that relied on it, as soon as more matrix obj types started to
always have `HasBaseDomain` set.

If one wants just a "suitable" field over which all matrices are represented,
one can use `DefaultScalarDomainOfMatrixList` (which perhaps should be renamed
or aliased to `DefaultBaseDomainOfMatrixList` ??). And in quite some cases,
maybe a completely different approach should be used...