Rename a local utility function
We rename `Phi2` to `Phi2_Md`. The function is only used locally for
the computation of certain numbers of conjugacy classes.

Background: A user reported a problem with using the data from
   https://www.math.rwth-aachen.de/~Frank.Luebeck/chev/DegMult/
in GAP4 which use variables Phi1, Phi2, and so on.

I do not request that we never use Phi1, and so on in GAP. But it seems
reasonable not to use Phi2 in this place.

(The mentioned data are currently referenced in 37 research articles in
MathSciNet. So, people might have code that uses these data such that I
prefer a solution without changing my data files.)