generate_release_notes.py: revise prioritylist
- add new section for HPC-GAP, gac, removal/deprecation
- fix typo: "kind: performance" -> "topic: performance"
- order more specific selectors before more general one:
  topic specific changes should come before general "kind" changes

Note that this is not necessarily the order in which we want to *show*
things in the release notes; but rather the order we use to decide in
which group we put everything. We may wish to make it possible to
control these two aspects independently.

Moreover, we could also change this to a multi-level model, where we
first group things by topic (with everything not in a specific recognize
topic put into a "General changes" group), and then for each group show
bugs/enhancements separately...