Fix `SaveOnExitFile` in restored workspaces
... and possibly more errors: we accidentally removed all exit handlers
when quitting GAP, just before the workspace was saved, thus when
restoring GAP from this session, those exist handlers were missing,
including the one dealing with `SaveOnExitFile`.

This code looses the ability to install new exit handlers as part of
an exit handler, but that seems like a rather bad (and pointless)
idea to me anyway.