Makefile.gappkg: ensure `make doc` terminates
Before this, `make doc` would end up in a GAP prompt, which is against
any convention and makes it unusable in automation.

Likewise, errors would just result in a break loop, instead of terminating
with a non-zero exit code.

This patch fixes this. It also suppresses the GAP banner for `make doc`.

To take effect, packages need to update to this new version.