Tweak ConvertToMatrixRepNC behavior, add tests
Before this, a command like

    ConvertToMatrixRepNC( [[1,0,1,1],[0,1,1,1]]*Z(2), 512 );

would result in an error. Now it returns `fail`, just like

    ConvertToMatrixRep( [[1,0,1,1],[0,1,1,1]]*Z(2), 512 );

does. The rationale here is that the request is reasonable
in principle; it only fails because we do not currently offer
compressed vectors over GF(512).

Contrast this with, say

    ConvertToMatrixRepNC( [[1,0,1,1],[0,1,1,1]]*Z(2), 3 );

where the inputs really do not make sense, so raising an error
here is justified (and we still do).

Also add some tests for the current behavior of `ConvertToMatrixRep` and
`ConvertToMatrixRepNC`; to accommodate them, I modified two `Assert`
statements in `ConvertToVectorRepNC`; one was redundant and removed, the
other was moved to a later point, to get a better error message in more
cases (when the error is raised before the assertion check is reached).