Unifying the references for matrix and vector
In the documentation, there were two reference names for the matrix: "mat" and "M". I unified it to "M" in every function, which was used from the beginning of the documentation. I did the same also for the vector which was referred to as "v" and "vec". I unified it to "v".