Makefile.gappkg: add simpler variant, remove comment
Remove a comment that does not match how packages currently use
Makefile.gappkg

Also add a much simpler variant of Makefile.gappkg that removes a bunch
of features (e.g. autoconf support, dependency tracking) and as a result
is much shorter and hopefully easier to understand. This is mostly
geared for simple packages with a single source file, as it also replace
the two step compilation process (.c/.cc -> .o -> .so) by direct
compilation (.c/.cc -> .so) which is fine if there is a single source
file and no header file.