fixed the method selection for constructors
Up to now, the installation of methods for constructors
does not take implied filters of the first argument into account.
This means that one can call the constructor with first argument
a filter that does not contain the filter from the method installation
but does contain the closure of that filter under implications.

(Think of a call of the constructor with first argument `IsMagma`.
A constructor method that "promises" an object in `IsGroup` shall be applicable.
Note that `TRUES_FLAGS( FLAGS_FILTER( IsGroup ) )` does not contain
`TRUES_FLAGS( FLAGS_FILTER( IsMagma ) )`,
but `TRUES_FLAGS( WITH_IMPS_FLAGS( FLAGS_FILTER( IsGroup ) ) )` does.)