fix the help system
If the Browse package is not loaded
and if the user preference `"Pager"` of GAP has the value `"builtin"`
and if a help request yields more than one match
then the entries of the list shown by the help request are not cut at
`SizeScreen()[1]` columns up to now.

For example, `?Random` yields several entries that need more than
80 columns to be shown.
Since the pager prints `SizeScreen()[2]` lines, the first entries
are not visible.

The current proposal fixes this.
Note that the GAPDoc function `InitialSubstringUTF8String` does not
take care about escape sequences.

Perhaps it would be more appropriate to add the new auxiliary function
`InitialSubstringUTF8Text` to the GAPDoc package.

(Besides this fix, the `\n` character at the end of the first line
in the pager has been removed, such that this line becomes visible.)