Fix a problem with kernel extensions on macOS
On macOS, kernel extensions built against non-installed GAP could
not be loaded into an executable built using libgap, such as
`tst/testlibgap/basic`. Resolve this by building kernel extensions
again with flat namespace (thus matching the behavior of the Linux
dynamic linker more closely), as we used to when we were still
using libtool for kernel extensions

This problem was only relevant for non-installed GAP builds.
For GAP installed via `make install` and kernel extensions built
against that, all was and is fine.