kernel: replace hack in the reader for setting function locks
The old code ignored the reader -> interpreter -> coder layering and
just assumed that the function currently being coded is referenced by
`CURR_FUNC()`. Instead we now inform the interpreter about the locks
which then (if applicable) informs the coder and finally that sets the
locks for the actual function being coded.