Simplify OneSM/OneAttr/OneOp/OneMutable/... etc.
Currently we have these three variants of `One`, together
with a bunch of synonyms for each:

- `OneImmutable` == `One` == `OneAttr` == `Identity`
- `OneMutable` == `OneOp`
- `OneSameMutability` == `OneSM`

This is then repeated for `Zero`, `Inverse`, `AdditiveInverse`.

The delegation between these is already complicated enough, but each
time I have to look into this, my head starts spinning due to the many
synonyms, and I can't quite remember which is which. I propose we
narrow this down a bit:

- the variants with `SM` seems to be purely to avoid some typing, so
  I've removed all uses of them. Turns out no packages uses it, either.
  So we can mark them as obsolete.

- the `Attr` variants are also almost unused, except for a single use in
  the FinInG package (removed in FinInG 1.5.4). Still, we can safely
  mark it as obsolete, and remove all users of it beyond the
  documentation. (As a bonus, also make `TransposedMatAttr` obsolete.)

- the `Op` synonym to me also seems redundant. Alas, a ton of packages
  use them, so I mostly left those alone.

- I think it is fine to keep `One` (and `Identity`) as synonyms for
  `OneImmutable`; at least for me personally it seems I can deal with
  that one synonym sufficiently well in my head.