Improve generate_package_release_notes.py
Now we pass in the old and new GAP version in as argument (instead of
hardcoding them). This is then used to download the appropriate package
infos.

Also fix a recently introduce bug in the script (were I accidentally
copy & pasted the variable name `new` instead of `pkg`), and fix an
older bug in `verify_via_checksumfile` (it was broken after a refactoring,
and nobody noticed because it only matters if SHA-256 validation fails...)