CI: revise the 'cygwin' job of the "Wrap releases" workflow
First, we change the 'unix' job to always upload the .tar.gz it produces
as an artifact. This puts a little extra burden on GitHub; but it lets
us modify the 'cygwin' job to always use this artifact.

This removes a major point in which our simulated releases differed from
actual releases. This subtle difference already caused trouble for two
releases. With this change, I hope it won't cause trouble a third time.

As an added benefit, it also allows us to remove a bunch of custom code
from the 'cygwin' job, and let's us avoid cloning GAP.